Add get user location endpoints

+++ routes/users.py
@@ -58,7 +58,8 @@ async def create_user(
     email: str = Form(...),
     password: str = Form(...),
     phone_number: str = Form(...),
-    user_profile:UploadFile=File(None)
+    user_profile:UploadFile=File(None),
+    location : str = Form("")
 ):
     """
     Create a new user.
@@ -90,16 +91,18 @@ async def create_user(
                 if result:
                     # Remove the temporary image file
                     profile_image_url = upload_file_to_drive(Path(f"temp/{user_profile.filename}"),"pubai")
-                    is_created:bool=User.create_user(session,username, email, password, phone_number,profile_image_url,id_,FERNET_KEY)
-                    
-                    os.remove(f"temp/{user_profile.filename}")
-                    return {"message": "User created successfully"}
+                    is_created:bool=User.create_user(session,username, email, password, phone_number,profile_image_url,id_,location,FERNET_KEY)
+                    if is_created:
+                        os.remove(f"temp/{user_profile.filename}")
+                        return {"message": "User created successfully"}
+                    else:
+                        raise UserExistException(f"User already exist with this email {email} or you have a problem in connexion")
             else:
                 os.remove(f"temp/{user_profile.filename}")
                 return HTTPException(status_code=400,detail="The image profile not respect the requirements")
         else:
             # Save user to database
-            is_created:bool=User.create_user(session,username, email, password, phone_number,None,None,FERNET_KEY)
+            is_created:bool=User.create_user(session,username, email, password, phone_number,None,None,location,FERNET_KEY)
         if not is_created:
             raise UserExistException(f"User already exist with this email {email} or you have a problem in connexion")
         
@@ -163,62 +166,28 @@ def get_user_by_access_token_(
     user:User|None=User.get_user_by_id(session,user_id)
     if user is None:
         raise HTTPException(status_code=401, detail="Invalid access token")
-    return {"username":user.username,"email":user.email,"phone_number":user.phone_number,"points":user.points,"profile_image":user.profile_image}
+    return {"username":user.username,"email":user.email,"phone_number":user.phone_number,
+            "points":user.points,"profile_image":user.profile_image,"location":user.location}
 
 
-
-@router.put("/points/add")
-def add_points(
-    points: int= Form(0),
+@router.put("/update-location")
+def update_location(
+    new_location: str= Form(...),
     access_token: str= Form(...)
     ):
-
-    # Check if points is less than zero to raise an appropriate error
-    if points < 0:
-        raise HTTPException(status_code=403, detail="Points cannot be negative")
     # Get the user from the database
-    decode_token:dict=decode_access_token(access_token,JWT_SECRET_KEY_ROBOT,ALGORITHM)
+    decode_token:dict=decode_access_token(access_token,JWT_SECRET_KEY,ALGORITHM)
     if not decode_token["valid"]:
         raise HTTPException(status_code=401, detail="Invalid access token")
     user_id:str=decode_token["user_id"]
     user:User|None=User.get_user_by_id(session,user_id)
     if user is None:
         raise HTTPException(status_code=401, detail="Invalid access token")
-    user.points+=points
-    # Set the new password
+    user.location = new_location
     session.commit()
     
     # Update the user in the database    
-    if points<10:
-        return {"message": f"{points} point added to {user.username} successfully"}
-    return {"message": f"{points} points added to {user.username} successfully"}
-
-@router.put("/points/deduct")
-def deduct_points(
-    points: int= Form(0),
-    access_token: str= Form(...)
-    ):
-    # Check if points is less than zero to raise an appropriate error
-    if points < 0:
-        raise HTTPException(status_code=403, detail="Points cannot be negative")
-    # Get the user from the database
-    decode_token:dict=decode_access_token(access_token,JWT_SECRET_KEY_ROBOT,ALGORITHM)
-    if not decode_token["valid"]:
-        raise HTTPException(status_code=401, detail="Invalid access token")
-    user_id:str=decode_token["user_id"]
-    user:User|None=User.get_user_by_id(session,user_id)
-    if user is None:
-        raise HTTPException(status_code=401, detail="Invalid access token")
-    if user.points<points:
-        raise HTTPException(status_code=401, detail="Insufficient points")
-    user.points-=points
-    # Set the new password
-    session.commit()
-    
-    # Update the user in the database    
-    if points<10:
-        return {"message": f"{points} point deducted from {user.username} successfully"}
-    return {"message": f"{points} points deducted from {user.username} successfully"}
+    return {"message": "Location changed successfully"}
